# Fix final encoding + update benches + perf