Perf looks good now, and we report the correct feedback when decoding the final quanta: 

```haskell
benchmarking decode/25/memory
time                 234.5 ns   (233.4 ns .. 235.5 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 234.1 ns   (233.0 ns .. 235.3 ns)
std dev              3.850 ns   (3.143 ns .. 4.788 ns)
variance introduced by outliers: 19% (moderately inflated)

benchmarking decode/25/base32
time                 55.86 ns   (55.63 ns .. 56.12 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 55.79 ns   (55.57 ns .. 56.11 ns)
std dev              855.5 ps   (675.8 ps .. 1.117 ns)
variance introduced by outliers: 19% (moderately inflated)

benchmarking decode/100/memory
time                 711.9 ns   (703.3 ns .. 719.0 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 701.6 ns   (697.3 ns .. 706.5 ns)
std dev              15.30 ns   (12.89 ns .. 18.86 ns)
variance introduced by outliers: 27% (moderately inflated)

benchmarking decode/100/base32
time                 169.0 ns   (167.7 ns .. 171.3 ns)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 172.3 ns   (170.9 ns .. 174.6 ns)
std dev              6.178 ns   (4.029 ns .. 10.65 ns)
variance introduced by outliers: 54% (severely inflated)

benchmarking decode/1k/memory
time                 6.254 μs   (6.203 μs .. 6.303 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 6.236 μs   (6.202 μs .. 6.276 μs)
std dev              124.4 ns   (101.1 ns .. 152.5 ns)
variance introduced by outliers: 20% (moderately inflated)

benchmarking decode/1k/base32
time                 1.503 μs   (1.491 μs .. 1.515 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.499 μs   (1.491 μs .. 1.509 μs)
std dev              31.67 ns   (26.73 ns .. 40.63 ns)
variance introduced by outliers: 25% (moderately inflated)

benchmarking decode/10k/memory
time                 61.94 μs   (61.52 μs .. 62.33 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 61.72 μs   (61.32 μs .. 62.16 μs)
std dev              1.412 μs   (1.141 μs .. 1.955 μs)
variance introduced by outliers: 19% (moderately inflated)

benchmarking decode/10k/base32
time                 14.94 μs   (14.78 μs .. 15.14 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 14.78 μs   (14.70 μs .. 14.89 μs)
std dev              320.7 ns   (261.0 ns .. 412.9 ns)
variance introduced by outliers: 21% (moderately inflated)

benchmarking decode/100k/memory
time                 638.2 μs   (633.0 μs .. 644.6 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 639.5 μs   (634.5 μs .. 646.1 μs)
std dev              20.00 μs   (14.97 μs .. 26.83 μs)
variance introduced by outliers: 23% (moderately inflated)

benchmarking decode/100k/base32
time                 148.7 μs   (146.9 μs .. 150.4 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 147.4 μs   (146.6 μs .. 148.5 μs)
std dev              3.362 μs   (2.570 μs .. 4.675 μs)
variance introduced by outliers: 17% (moderately inflated)

benchmarking decode/1mm/memory
time                 6.251 ms   (6.205 ms .. 6.299 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 6.299 ms   (6.256 ms .. 6.347 ms)
std dev              132.2 μs   (107.9 μs .. 170.2 μs)

benchmarking decode/1mm/base32
time                 1.526 ms   (1.494 ms .. 1.554 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 1.484 ms   (1.474 ms .. 1.499 ms)
std dev              43.67 μs   (33.99 μs .. 56.89 μs)
variance introduced by outliers: 17% (moderately inflated)
```